+++ pkg/apis/training/v1alpha1/notebook_types.go
@@ -105,7 +105,7 @@ type NotebookSpec struct {
 	// ActiveDeadlineSeconds is the deadline of a job for this notebook.
 	// +kubebuilder:default:=600
 	// +kubebuilder:validation:Optional
-	ActiveDeadlineSeconds *int32 `json:"activeDeadlineSeconds,omitempty" protobuf:"varint,11,opt,name=activeDeadlineSeconds"`
+	ActiveDeadlineSeconds *int64 `json:"activeDeadlineSeconds,omitempty" protobuf:"varint,11,opt,name=activeDeadlineSeconds"`
 }
 
 // NotebookStatus is the observed state of the notebook resource
